Declare support for the pixel formats: PIX_FMT_YUVJ420P PIX_FMT_YUVJ422P PIX_FMT_YUVJ440P PIX_FMT_YUVJ444P
in the isSupported{In,Out} macros. 

These pixel formats are not true pixel formats but hacks specific to
JPEG in libavcodec. They are deprecated and should be removed (that is
from libavcodec first and libswscale second)...  but they must be
tested by swscale-test.

See thread:
Subject: [FFmpeg-devel] [PATCH] Extend show_pix_fmts() to make it print the input/output support
Date: 2010-01-30 15:54:08 GMT


git-svn-id: file:///var/local/repositories/mplayer/trunk/libswscale@30474 b3059339-0415-0410-9bf9-f77b7e298cf2